+++ api-v2/conf/routes
@@ -4,4 +4,4 @@ GET    /readiness    com.foreignlanguagereader.api.controller.v1.HealthControlle
 
 GET    /v1/language/definition/:wordlanguage/:word/ com.foreignlanguagereader.api.controller.v1.language.DefinitionController.definition(wordlanguage: com.foreignlanguagereader.api.domain.Language.Language, word: String)
 
-POST  /graphql  controllers.GraphQL.graphql
+POST  /graphql  com.foreignlanguagereader.api.controller.v1.graphql.GraphQL.graphql
